Moved definition to cc file for external linkage
